chore: add category as separate section
