+++ apps/meteor/client/navbar/NavBarPagesGroup/hooks/useCreateNewMenu.ts
@@ -2,6 +2,7 @@ import { useAtLeastOnePermission } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
 import { useCreateNewItems } from './useCreateNewItems';
+import useCreateCategoryModal from '../../../sidebar/categories/useCreateCategoryModal';
 
 const CREATE_ROOM_PERMISSIONS = ['create-c', 'create-p', 'create-d', 'start-discussion', 'start-discussion-other-user'];
 
@@ -10,8 +11,12 @@ export const useCreateNewMenu = () => {
 	const showCreate = useAtLeastOnePermission(CREATE_ROOM_PERMISSIONS);
 
 	const createRoomItems = useCreateNewItems();
+	const createCategoriesItem = useCreateCategoryModal();
 
-	const sections = [{ title: t('Create_new'), items: createRoomItems, permission: showCreate }];
+	const sections = [
+		{ title: t('Create_new'), items: createRoomItems, permission: showCreate },
+		{ title: '', items: [createCategoriesItem], permission: true },
+	];
 
 	return sections.filter((section) => section.permission);
 };
